requirements.txt
certifi==2020.12.5
chardet==3.0.4
cheroot==8.5.0
cssselect==1.1.0
idna==2.10
install==1.3.4
jaraco.functools==3.0.1
lxml==4.6.2
more-itertools==8.6.0
pypandoc==1.5
readability-lxml==0.8.1
requests==2.25.0
six==1.15.0
urllib3==1.26.2
web.py==0.62
